Add defer to main.go
Add function defer to main.go. If a panic occurs this should be the last defered
function. This function catches a panic, and if there is an exit instruction in it
will exit instead. Effectively this allows all deffered functions to complete before exiting, and prevents a
staktrace from printing. We want to use panic instead of fatal, but for expected exits we don't need the stacktrace