Improve readCSV error message
Vendcli is used by persons with not great cli experience. A common error they run into is not
correctly specifying the path of a file they'd like to use with vendcli. The error message when
they don't choose the correct path is generic and leads to confusion. This commit improves the
error message and makes more verbose, and gives a tip to beginers

can be combined