Add support for multiple suppliers to export-product command
Retailers have requested a way of viewing the supplier information per product on each line. added this so each supplier will
have its own column. This change has vendcli loop through the produce slices first, and grab the max number of suppliers a
given product will have. it uses this max number to determine the number of columns to make