Fixing SKU order and making headers pretty
SKU Types are not always in order. This change adds a sorting portion to the helper functions so that both the making of the headers and records maintain the same order.

Old versions of sku codes also are in lower case. This change adds a strings.ToUpper portion to maintain the same UPPER for all code Types. This is something we also have to do in the DB when querying for secondary codes.

This change also removes the index number from the first version of each SKU type to clean up the CSV file a bit.